+++ CHANGES.md
@@ -6,6 +6,8 @@
 - Added the option of making API routes public in the fastapi Resource.
   This is done by moving the auth dependencies from app to route level.
 
+- Bumped fastapi to 0.110.* and pyjwt to 2.8.*.
+
 
 ## 0.12.3 (2024-03-20)
 -------------------